// src/template/state/type.ts
export enum Type {
  SetPageSizeWidth = "SET_PAGE_SIZE_WIDTH",
  SetPageSizeHeight = "SET_PAGE_SIZE_HEIGHT",
  SetPageMarginLeft = "SET_PAGE_MARGIN_LEFT",
  SetPageMarginTop = "SET_PAGE_MARGIN_TOP",
  SetPageMarginRight = "SET_PAGE_MARGIN_RIGHT",
  SetPageMarginBottom = "SET_PAGE_MARGIN_BOTTOM",
  AddTemplateItem = "ADD_TEMPLATE_ITEM",
  RemoveTemplateItem = "REMOVE_TEMPLATE_ITEM",
  SetTemplateItemValue = "SET_TEMPLATE_ITEM_VALUE"
}